Cast Twitter Labs API results as json
The new Labs route do not end with .json but their results are still json. So we have to set `format=""` to properly call the API, but in such case the results are not cast as JSON, so this hardfixes it by always casting results of calls to LABS as JSON

But since there is to my knowledge no more Twitter route returning any other format than JSON, I guess we could rather always cast the results as JSON, and maybe remove completely the `format` argument from the class and replace it with a suffix one, which could be always forced to .json when `api_version == 1.1`

What do you think @sixohsix?